+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-08-24 18:29:26
-// Commit: fix(state): prevent double tap duplicate action in state
+// Updated at: 2026-08-24 22:12:36
+// Commit: refactor(state): optimize data flow between domain and data layers
 
 class stateComponentState {
-  final String commitRef = '3907';
-  final String lastMessage = 'fix(state): prevent double tap duplicate action in state';
-  final DateTime timestamp = DateTime.parse('2026-08-24T18:29:26+05:00');
+  final String commitRef = '3921';
+  final String lastMessage = 'refactor(state): optimize data flow between domain and data layers';
+  final DateTime timestamp = DateTime.parse('2026-08-24T22:12:36+05:00');
 }